+++ packages/sdk/test/unit/GroupKeyManager.test.ts
@@ -8,7 +8,6 @@ import { createPrivateKeyAuthentication } from '../../src/Authentication'
 import { DestroySignal } from '../../src/DestroySignal'
 import { GroupKey } from '../../src/encryption/GroupKey'
 import { GroupKeyManager } from '../../src/encryption/GroupKeyManager'
-import { LitProtocolFacade } from '../../src/encryption/LitProtocolFacade'
 import { LocalGroupKeyStore } from '../../src/encryption/LocalGroupKeyStore'
 import { SubscriberKeyExchange } from '../../src/encryption/SubscriberKeyExchange'
 import { StreamrClientEventEmitter } from '../../src/events'
@@ -19,7 +18,6 @@ const GROUP_KEY = GroupKey.generate('groupKeyId-123')
 describe('GroupKeyManager', () => {
 
     let groupKeyStore: MockProxy<LocalGroupKeyStore>
-    let litProtocolFacade: MockProxy<LitProtocolFacade>
     let subscriberKeyExchange: MockProxy<SubscriberKeyExchange>
     let eventEmitter: StreamrClientEventEmitter
     let groupKeyManager: GroupKeyManager
@@ -27,15 +25,12 @@ describe('GroupKeyManager', () => {
 
     const getUserId = () => toUserId(wallet.address)
 
-    function createGroupKeyManager(litProtocolEnabled: boolean): GroupKeyManager {
+    function createGroupKeyManager(): GroupKeyManager {
         return new GroupKeyManager(
             subscriberKeyExchange,
-            litProtocolFacade,
             groupKeyStore,
             {
                 encryption: {
-                    litProtocolEnabled,
-                    litProtocolLogging: false,
                     maxKeyRequestsPerSecond: 10,
                     keyRequestTimeout: 100
                 } as any
@@ -52,10 +47,9 @@ describe('GroupKeyManager', () => {
 
     beforeEach(() => {
         groupKeyStore = mock<LocalGroupKeyStore>()
-        litProtocolFacade = mock<LitProtocolFacade>()
         subscriberKeyExchange = mock<SubscriberKeyExchange>()
         eventEmitter = new StreamrClientEventEmitter()
-        groupKeyManager = createGroupKeyManager(true)
+        groupKeyManager = createGroupKeyManager()
     })
 
     describe('fetchKey', () => {
@@ -66,18 +60,6 @@ describe('GroupKeyManager', () => {
             expect(key).toEqual(GROUP_KEY)
             expect(groupKeyStore.get).toHaveBeenCalledWith(GROUP_KEY.id, getUserId())
             expect(groupKeyStore.get).toHaveBeenCalledTimes(1)
-            expect(litProtocolFacade.get).toHaveBeenCalledTimes(0)
-            expect(subscriberKeyExchange.requestGroupKey).toHaveBeenCalledTimes(0)
-        })
-
-        it('key present in lit protocol', async () => {
-            litProtocolFacade.get.mockResolvedValueOnce(GROUP_KEY)
-
-            const key = await groupKeyManager.fetchKey(toStreamPartID(STREAM_ID, 0), GROUP_KEY.id, getUserId())
-            expect(key).toEqual(GROUP_KEY)
-            expect(groupKeyStore.get).toHaveBeenCalledTimes(1)
-            expect(litProtocolFacade.get).toHaveBeenCalledWith(STREAM_ID, GROUP_KEY.id)
-            expect(litProtocolFacade.get).toHaveBeenCalledTimes(1)
             expect(subscriberKeyExchange.requestGroupKey).toHaveBeenCalledTimes(0)
         })
 
@@ -90,7 +72,6 @@ describe('GroupKeyManager', () => {
             const key = await groupKeyManager.fetchKey(toStreamPartID(STREAM_ID, 0), GROUP_KEY.id, getUserId())
             expect(key).toEqual(GROUP_KEY)
             expect(groupKeyStore.get).toHaveBeenCalledTimes(2)
-            expect(litProtocolFacade.get).toHaveBeenCalledTimes(1)
             expect(subscriberKeyExchange.requestGroupKey).toHaveBeenCalledWith(
                 GROUP_KEY.id,
                 getUserId(),
@@ -104,52 +85,21 @@ describe('GroupKeyManager', () => {
                 .rejects
                 .toThrow('waitForEvent (timed out after 100 ms)')
             expect(groupKeyStore.get).toHaveBeenCalledTimes(1)
-            expect(litProtocolFacade.get).toHaveBeenCalledTimes(1)
-            expect(subscriberKeyExchange.requestGroupKey).toHaveBeenCalledTimes(1)
-        })
-
-        it('skips lit protocol if lit protocol disabled in config', async () => {
-            groupKeyManager = createGroupKeyManager(false)
-            await expect(groupKeyManager.fetchKey(toStreamPartID(STREAM_ID, 0), GROUP_KEY.id, getUserId()))
-                .rejects
-                .toThrow('waitForEvent (timed out after 100 ms)')
-            expect(groupKeyStore.get).toHaveBeenCalledTimes(1)
-            expect(litProtocolFacade.get).toHaveBeenCalledTimes(0)
             expect(subscriberKeyExchange.requestGroupKey).toHaveBeenCalledTimes(1)
         })
     })
 
     describe('storeKey', () => {
-        it('given pre-defined key only stores in (local) group key store and skips lit protocol', async () => {
+        it('given pre-defined key stores in (local) group key store', async () => {
             const returnedGroupKey = await groupKeyManager.storeKey(GROUP_KEY, getUserId(), STREAM_ID)
             expect(returnedGroupKey).toEqual(GROUP_KEY)
             expect(groupKeyStore.set).toHaveBeenCalledWith(GROUP_KEY.id, getUserId(), GROUP_KEY.data)
-            expect(litProtocolFacade.store).toHaveBeenCalledTimes(0)
         })
 
         describe('given no pre-defined key', () => {
-            it('lit-protocol online: fetches key from lit protocol and then stores in (local) group key as well', async () => {
-                litProtocolFacade.store.mockImplementationOnce(async (_streamId, symmetricKey) => {
-                    return new GroupKey('foobarId', Buffer.from(symmetricKey))
-                })
-
-                const returnedGroupKey = await groupKeyManager.storeKey(undefined, getUserId(), STREAM_ID)
-                expect(returnedGroupKey.id).toEqual('foobarId')
-                expect(groupKeyStore.set).toHaveBeenCalledWith(returnedGroupKey.id, getUserId(), returnedGroupKey.data)
-                expect(litProtocolFacade.store).toHaveBeenCalledWith(STREAM_ID, returnedGroupKey.data)
-
-            })
-
-            it('lit-protocol offline: generates new key and stores only in (local) group key store', async () => {
+            it('generates new key and stores only in (local) group key store', async () => {
                 const returnedGroupKey = await groupKeyManager.storeKey(undefined, getUserId(), STREAM_ID)
                 expect(groupKeyStore.set).toHaveBeenCalledWith(returnedGroupKey.id, getUserId(), returnedGroupKey.data)
-                expect(litProtocolFacade.store).toHaveBeenCalledWith(STREAM_ID, returnedGroupKey.data)
-            })
-
-            it('lit-protocol disabled: does not even attempt to store key in lit protocol', async () => {
-                groupKeyManager = createGroupKeyManager(false)
-                await groupKeyManager.storeKey(undefined, getUserId(), STREAM_ID)
-                expect(litProtocolFacade.store).toHaveBeenCalledTimes(0)
             })
         })
     })