+++ packages/trackerless-network/test/unit/HandshakeRpcLocal.test.ts
@@ -1,4 +1,4 @@
-import { DhtAddress, NodeType, toNodeId, toDhtAddressRaw } from '@streamr/dht'
+import { DhtAddress, toNodeId } from '@streamr/dht'
 import { StreamPartIDUtils } from '@streamr/utils'
 import { NodeList } from '../../src/logic/NodeList'
 import { HandshakeRpcLocal } from '../../src/logic/neighbor-discovery/HandshakeRpcLocal'
@@ -85,10 +85,7 @@ describe('HandshakeRpcLocal', () => {
 
     it('handshakeWithInterleaving success', async () => {
         const req: InterleaveRequest = {
-            interleaveTargetDescriptor: {
-                nodeId: toDhtAddressRaw('0x2222' as DhtAddress),
-                type: NodeType.NODEJS
-            }
+            interleaveTargetDescriptor: createMockPeerDescriptor()
         }
         await rpcLocal.interleaveRequest(req, {
             incomingSourceDescriptor: createMockPeerDescriptor()