+++ packages/trackerless-network/test/unit/NodeList.test.ts
@@ -1,6 +1,5 @@
 import {
     ListeningRpcCommunicator,
-    NodeType,
     PeerDescriptor,
     randomDhtAddress,
     toDhtAddress,
@@ -42,27 +41,20 @@ describe('NodeList', () => {
     beforeEach(() => {
         nodeList = new NodeList(ownId, 6)
         for (const id of ids) {
-            const peerDescriptor: PeerDescriptor = {
-                nodeId: id,
-                type: NodeType.NODEJS
-            }
+            const peerDescriptor = createMockPeerDescriptor({
+                nodeId: id
+            })
             nodeList.add(createRemoteGraphNode(peerDescriptor))
         }
     })
 
     it('add', () => {
-        const newDescriptor = {
-            nodeId: new Uint8Array([1, 2, 3]),
-            type: NodeType.NODEJS
-        }
+        const newDescriptor = createMockPeerDescriptor()
         const newNode = createRemoteGraphNode(newDescriptor)
         nodeList.add(newNode)
         expect(nodeList.has(toNodeId(newDescriptor))).toEqual(true)
 
-        const newDescriptor2 = {
-            nodeId: new Uint8Array([1, 2, 4]),
-            type: NodeType.NODEJS
-        }
+        const newDescriptor2 = createMockPeerDescriptor()
         const newNode2 = createRemoteGraphNode(newDescriptor2)
         nodeList.add(newNode2)
         expect(nodeList.has(toNodeId(newDescriptor2))).toEqual(false)